third_party/WebKit/LayoutTests/fast/css/usecounter-negative-mask-size-shorthand.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
'use strict';

test(() => {
    let NegativeMaskSize = 2403; // From web_feature.mojom

    let isCounted = () => internals.isUseCounted(document, NegativeMaskSize);
    var div = document.createElement('div');

    div.style = '-webkit-mask: -3px -4px / 1px 2px;';
    div.style = 'background: none, red none 50% 50% / -1px -2px round space local padding-box content-box;';
    assert_false(isCounted(),
                 'background should not be counted');

    div.style = '-webkit-mask: 3px 4px / -1px 2px;';
    assert_true(isCounted(),
                '-webkit-mask should be counted');
}, 'Negative size is use counted for -webkit-mask-size in shorthand');
</script>
